add query handling and create smaller code chunks
